chore: move GrammarParser -> Grammar/Parser
