+++ tests/Unit/GrammarParserTest.php
@@ -1,14 +1,14 @@
 <?php
 
-use Phiki\GrammarParser;
+use Phiki\Grammar\Parser;
 
 describe('GrammarParser', function () {
     it('can be constructed', function () {
-        expect(new GrammarParser)->toBeInstanceOf(GrammarParser::class);
+        expect(new Parser)->toBeInstanceOf(Parser::class);
     });
 
     it('can parse a grammar file', function () {
-        $parser = new GrammarParser;
+        $parser = new Parser;
         $grammar = $parser->parse(json_decode(file_get_contents(__DIR__.'/../../languages/php.json'), true));
 
         expect($grammar->scopeName)->toBe('source.php');
@@ -17,7 +17,7 @@ describe('GrammarParser', function () {
     });
 
     it('can parse a grammar file with injections', function () {
-        $parser = new GrammarParser;
+        $parser = new Parser;
         $grammar = $parser->parse(json_decode(file_get_contents(__DIR__.'/../../languages/blade.json'), true));
 
         expect($grammar->scopeName)->toBe('text.html.php.blade');
@@ -26,7 +26,7 @@ describe('GrammarParser', function () {
     });
 
     it('marks injection patterns as injected ones', function () {
-        $parser = new GrammarParser;
+        $parser = new Parser;
         $grammar = $parser->parse(json_decode(file_get_contents(__DIR__.'/../../languages/blade.json'), true));
 
         expect($grammar->scopeName)->toBe('text.html.php.blade');
